chore(test): extract pose.rs test module into pose/tests.rs
Same maintainability split as tui_renderer: pose.rs was ~2490 lines, ~1800 of
them tests. Now pose/mod.rs (production: routed pose, snap-back, freeze) +
pose/tests.rs (unit + frame-by-frame continuity tests) as a
`#[cfg(test)] mod tests;` submodule that keeps `use super::*` and full
crate-internal access — no API change. Production pose logic is now ~680
readable lines.

561 tests, clippy -D warnings + fmt clean.
